+++ src/cmd/mod.rs
@@ -16,6 +16,8 @@ mod doctor;
 mod info;
 pub(crate) mod install;
 mod list;
+#[cfg(feature = "lsp")]
+pub(crate) mod lsp;
 #[cfg(feature = "man")]
 mod man;
 pub(crate) mod run;
@@ -33,6 +35,8 @@ pub(crate) use list::list;
 #[cfg(feature = "man")]
 pub(crate) use man::{write_man_pages, write_runner_page_to_stdout};
 pub(crate) use run::run;
+#[cfg(feature = "schema")]
+pub(crate) use schema::config_schema;
 #[cfg(feature = "schema")]
 pub(crate) use schema::write_schema;
 pub(crate) use why::why;